https://github.com/mP1/walkingkooka-spreadsheet/pull/7203 viewport package extracted
- https://github.com/mP1/walkingkooka-spreadsheet/pull/7203
- viewport package extracted
