+++ src/test/java/walkingkooka/spreadsheet/server/meta/BasicSpreadsheetMetadataHateosResourceHandlerContextTest.java
@@ -68,7 +68,6 @@ import walkingkooka.spreadsheet.provider.SpreadsheetProviders;
 import walkingkooka.spreadsheet.reference.SpreadsheetCellReference;
 import walkingkooka.spreadsheet.reference.SpreadsheetLabelName;
 import walkingkooka.spreadsheet.reference.SpreadsheetSelection;
-import walkingkooka.spreadsheet.reference.SpreadsheetViewport;
 import walkingkooka.spreadsheet.security.store.SpreadsheetGroupStores;
 import walkingkooka.spreadsheet.security.store.SpreadsheetUserStores;
 import walkingkooka.spreadsheet.store.SpreadsheetCellRangeStores;
@@ -81,6 +80,7 @@ import walkingkooka.spreadsheet.store.SpreadsheetRowStores;
 import walkingkooka.spreadsheet.store.repo.SpreadsheetStoreRepositories;
 import walkingkooka.spreadsheet.store.repo.SpreadsheetStoreRepository;
 import walkingkooka.spreadsheet.validation.form.store.SpreadsheetFormStores;
+import walkingkooka.spreadsheet.viewport.SpreadsheetViewport;
 import walkingkooka.storage.StorageStores;
 import walkingkooka.store.Store;
 import walkingkooka.tree.expression.function.provider.ExpressionFunctionInfoSet;
